+++ src/core/index.ts
@@ -6,6 +6,7 @@ import { YAMLConverter } from "./converters/yaml";
 import { TOMLConverter } from "./converters/toml";
 import { ENVConverter } from "./converters/env";
 import { XMLConverter } from "./converters/xml";
+import { INIConverter } from "./converters/ini";
 import {
   ConfigFormat,
   ConfigData,
@@ -31,6 +32,7 @@ export class UniversalConfigConverter {
     this.registry.register(new TOMLConverter());
     this.registry.register(new ENVConverter());
     this.registry.register(new XMLConverter());
+    this.registry.register(new INIConverter());
   }
 
   /**
@@ -156,4 +158,5 @@ export {
   TOMLConverter,
   ENVConverter,
   XMLConverter,
+  INIConverter,
 };